Add 10 - color keying
This one is more deliberating as it's not just a single C source file
anymore. Going in with non-OOP appraoch in pure C is fun.

Notice 'extern' is used in source file that refers and uses variables
defined externally. It's not required but this will make it more clear
(the same way of 'extern' in front of function declaration which is not
required but it's there by default.).